Delete functions, correct VTT format

+++ app/Http/Controllers/Api/VttController.php
@@ -4,11 +4,8 @@ namespace App\Http\Controllers\Api;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
-use Carbon\Carbon;
-use DateTime;
 use Illuminate\Support\Facades\Http;
 use Jwplayer\JwplatformAPI;
-use Illuminate\Support\Facades\File;
 use App\Classes\vttConstructor;
 use Podlove\Webvtt\Parser;
 
@@ -18,26 +15,28 @@ class VttController extends Controller
 {
     public function UploadCaption(Request $request)
     {
-         $validateData = $request->validate([
-             'VttData' => 'required',
-             'video_key' => 'required',
-             'kind' => 'required',
-             'label' => 'required',
-         ]);
-        $completeVttString = vttConstructor::constructVtt($validateData['VttData']);
-        $id =  str_replace(".", "", uniqid( "", true));
-        $fp = fopen(storage_path( "app/tempfiles/" . $id . ".vtt"), "wb");
+        $validateData = $request->validate([
+            'VttData' => 'required',
+            'video_key' => 'required',
+            'kind' => 'required',
+            'label' => 'required',
+            'language' => 'required',
+        ]);
+        $completeVttString = vttConstructor::constructVtt($validateData['VttData'], $validateData['kind'], $validateData['language']);
+        $id =  str_replace(".", "", uniqid("", true));
+        $fp = fopen(storage_path("app/tempfiles/" . $id . ".vtt"), "wb");
         fwrite($fp, $completeVttString);
         fclose($fp);
         $jwplatform_api = new JwplatformAPI(env("JWPLAYER_API_KEY", ""), env("JWPLAYER_API_SECRET", ""));
         $params = [
-                    'video_key' => $validateData['video_key'],
-                    'kind' => $validateData['kind'],
-                    'label' => $validateData['label']
-                ];
+            'video_key' => $validateData['video_key'],
+            'kind' => $validateData['kind'],
+            'label' => $validateData['label'],
+            'language' => $validateData['language']
+        ];
         $url = "app/tempfiles/" . $id . ".vtt";
         $decoded = json_decode(trim(json_encode($jwplatform_api->call('/videos/tracks/create', $params))), TRUE);
-        $upload_link = $decoded ['link'];
+        $upload_link = $decoded['link'];
         $upload_response = $jwplatform_api->upload(storage_path($url), $upload_link);
         unlink(storage_path($url));
         return response(['message' => $upload_response["status"]], 200);
@@ -55,7 +54,6 @@ class VttController extends Controller
 
         return $result != null ? response(['VttData' => $result], 200) : response(['error' => 'Internal server error'], 500);
     }
-    
 
     public function SaveCaption(Request $request)
     {
@@ -65,28 +63,49 @@ class VttController extends Controller
             'label' => 'required',
         ]);
         $explodedLink = explode("/", $validateData['VttLink']);
-        $trackKey = str_replace(".tmp", "",str_replace(".vtt", "", $explodedLink[count($explodedLink) - 1]));
+        $trackKey = str_replace(".tmp", "", str_replace(".vtt", "", $explodedLink[count($explodedLink) - 1]));
         $completeVttString = vttConstructor::constructVtt($validateData['VttData']);
-        $id =  str_replace(".", "", uniqid( "", true));
-        $fp = fopen(storage_path( "app/tempfiles/" . $id . ".vtt"), "wb");
+        $id =  str_replace(".", "", uniqid("", true));
+        $fp = fopen(storage_path("app/tempfiles/" . $id . ".vtt"), "wb");
         fwrite($fp, $completeVttString);
         fclose($fp);
         $jwplatform_api = new JwplatformAPI(env("JWPLAYER_API_KEY", ""), env("JWPLAYER_API_SECRET", ""));
         $params = [
-                    'track_key' => $trackKey,
-                    'label' => $validateData['label'],
-                    'update_file' => 'true',
-                ];
+            'track_key' => $trackKey,
+            'label' => $validateData['label'],
+            'update_file' => 'true',
+        ];
         $url = "app/tempfiles/" . $id . ".vtt";
         $decoded = json_decode(trim(json_encode($jwplatform_api->call('/videos/tracks/update', $params))), TRUE);
-        $upload_link = $decoded ['link'];
-        $upload_response = $jwplatform_api->upload(storage_path($url), $upload_link);
-        unlink(storage_path($url));
-        return response(['message' => $upload_response["status"]], 200);
+        if (!array_key_exists('message', $decoded)) {
+            $upload_link = $decoded['link'];
+            $upload_response = $jwplatform_api->upload(storage_path($url), $upload_link);
+            unlink(storage_path($url));
+            return response(['message' => $upload_response["status"]], 200);
+        }
 
+        return response(['error' => $decoded["message"]], 500);
     }
 
     public function DeleteCaption(Request $request)
     {
+        $validateData = $request->validate([
+            'VttLink' => 'required',
+        ]);
+        $explodedLink = explode("/", $validateData['VttLink']);
+        $trackKey = str_replace(".tmp", "", str_replace(".vtt", "", $explodedLink[count($explodedLink) - 1]));
+        $jwplatform_api = new JwplatformAPI(env("JWPLAYER_API_KEY", ""), env("JWPLAYER_API_SECRET", ""));
+        $params = [
+            'track_key' => $trackKey,
+        ];
+        $decoded = json_decode(trim(json_encode($jwplatform_api->call('/videos/tracks/delete', $params))), TRUE);
+        
+        if($decoded['status'] == 'ok'){
+            return response(['status' => $decoded['status']], 200);
+        }
+        else if(array_key_exists('message', $decoded)){
+            return response(['message' => $decoded['message']], 500);
+        }
+        
     }
 }
